:memo: Add and update docs for using the request body in exception handlers
